Update install.sh
apply patch only when ubuntu >= 24.04

+++ util/install.sh
@@ -137,8 +137,10 @@ function mn_deps {
 
     sudo git clone https://github.com/mininet/mininet.git
     pushd $MININET_DIR/mininet-wifi/mininet
-    git reset --hard 6eb8973
-    patch -p0 < $MININET_DIR/mininet-wifi/util/mininet-patches/mininet.patch
+    if [ "$DIST" = "Ubuntu" ] &&  [ `expr $RELEASE '>=' 24.04` = "1" ]; then
+        git reset --hard 6eb8973
+        patch -p0 < $MININET_DIR/mininet-wifi/util/mininet-patches/mininet.patch
+    fi
     sudo PYTHON=${PYTHON} make install
     popd
     echo "Installing Mininet-wifi core"
